requirements.txt
pywebview>=4.0.0
pyinstaller>=5.0.0
tomli>=2.0.0;python_version<'3.11'
# Windows-specific: Edge WebView2 runtime (installed separately)
